# ENH: stats: override _rvs for nhypergeom