#### Reference issue
gh-10106

#### What does this implement/fix?
Overrides `_rvs` for `nhypergeom`, which was the slowest among the discrete distributions.

The approach is to create a fast inverse cdf with `interp1d` then do inverse transform sampling. The support of the distribution is the integers `0, 1, ..., n)` so this tends to be quite fast, and it's exactly as accurate as the `cdf`.

My guess is that `ppf` is more often called for one or a few points whereas it is common for many random variates to be generated at once, so I put the cdf inversion inside `_rvs` only rather than in `_ppf`.

This could be further improved, of course (e.g. use the default if `c * np.prod(size) < n`, where `c` is the expected number of `cdf` evaluations needed for the default `ppf` method to converge), but I'm not sure if that's important.

#### Additional information
~~Mostly I did this to better understand how `rvs` works with `ndarray` shape parameters. If this approach is OK, I will try to break the approach into a function so that there is an easy, standard way of getting the desired broadcasting behavior from an _rvs function that natively handles only scalar shape parameters.~~

_Update: now uses the decorator written for the `nchypergeom_...`s to do the broadcasting._