@mdhaber 2021-02-26
@tirthasheshpatel What do you think of this as a way to speed up RVS for `nhypergeom`?

```python3
from scipy.stats import nhypergeom 
%timeit nhypergeom.rvs(22, 7, 13, size=10000)
```
Before:
```
4.11 s ± 7.84 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
```
After:
```
1.12 ms ± 1.98 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)
```
Identical random variates.