@tirthasheshpatel 2021-02-27
> @tirthasheshpatel What do you think of this as a way to speed up RVS for `nhypergeom`?

Thanks, @mdhaber. It's really cool. Simple and fast!